% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/vertcloseness-package.R
\docType{package}
\name{closeness_by_name}
\arguments{
  \item{vertice_name}
  
\alias{vertcloseness-package}
\title{vertice closeness}
\description{
packague tht give the closeness metric toa given vertice}

