% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/vertcloseness-package.R
\docType{package}
\name{vertice closeness}
\arguments{
  \item{vertice_name}{your name. Required.}
  
\alias{vertcloseness-package}
\title{vertice closeness}
\description{
Describe the package.
}

